[android] resize Android launcher and notification icons during build/detach
